fix: Replace integration tests with simple unit tests to avoid CI API key issues
- Remove problematic integration tests that require real config loading
- Add simple unit tests that properly mock all dependencies
- Tests now run without requiring OPENROUTER_API_KEY environment variable
